Move show_commit module to core
