+++ src/core/modules/mod.rs
@@ -4,6 +4,7 @@ mod error;
 mod external_editor;
 mod insert;
 mod list;
+mod show_commit;
 
 pub use self::{
 	confirm_abort::ConfirmAbort,
@@ -12,4 +13,5 @@ pub use self::{
 	external_editor::ExternalEditor,
 	insert::Insert,
 	list::List,
+	show_commit::ShowCommit,
 };
