+++ src/core/modules/show_commit/file_stats_builder.rs
@@ -1,4 +1,4 @@
-use crate::show_commit::{delta::Delta, diff_line::DiffLine, file_stat::FileStat};
+use super::{delta::Delta, diff_line::DiffLine, file_stat::FileStat};
 
 #[derive(Debug, Clone)]
 pub(super) struct FileStatsBuilder {
@@ -60,7 +60,7 @@ impl FileStatsBuilder {
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use crate::show_commit::{origin::Origin, status::Status};
+	use crate::core::modules::show_commit::{origin::Origin, status::Status};
 
 	#[test]
 	fn build_file_stat_with_file_stat_without_delta() {
